Fix SWR key logic
