+++ resources/scripts/api/account/activity.ts
@@ -1,10 +1,10 @@
-import useUserSWRContentKey from '@/plugins/useUserSWRContentKey';
 import useSWR, { ConfigInterface, responseInterface } from 'swr';
 import { ActivityLog, Transformers } from '@definitions/user';
 import { AxiosError } from 'axios';
 import http, { PaginatedResult, QueryBuilderParams, withQueryBuilderParams } from '@/api/http';
 import { toPaginatedSet } from '@definitions/helpers';
 import useFilteredObject from '@/plugins/useFilteredObject';
+import { useUserSWRKey } from '@/plugins/useSWRKey';
 
 export type ActivityLogFilters = QueryBuilderParams<'ip' | 'event', 'timestamp'>;
 
@@ -12,7 +12,7 @@ const useActivityLogs = (
     filters?: ActivityLogFilters,
     config?: ConfigInterface<PaginatedResult<ActivityLog>, AxiosError>
 ): responseInterface<PaginatedResult<ActivityLog>, AxiosError> => {
-    const key = useUserSWRContentKey(['account', 'activity', JSON.stringify(useFilteredObject(filters || {}))]);
+    const key = useUserSWRKey(['account', 'activity', JSON.stringify(useFilteredObject(filters || {}))]);
 
     return useSWR<PaginatedResult<ActivityLog>>(
         key,
