+++ resources/scripts/api/server/activity.ts
@@ -4,6 +4,7 @@ import { AxiosError } from 'axios';
 import http, { PaginatedResult, QueryBuilderParams, withQueryBuilderParams } from '@/api/http';
 import { toPaginatedSet } from '@definitions/helpers';
 import useFilteredObject from '@/plugins/useFilteredObject';
+import { useServerSWRKey } from '@/plugins/useSWRKey';
 import { ServerContext } from '@/state/server';
 
 export type ActivityLogFilters = QueryBuilderParams<'ip' | 'event', 'timestamp'>;
@@ -12,10 +13,11 @@ const useActivityLogs = (
     filters?: ActivityLogFilters,
     config?: ConfigInterface<PaginatedResult<ActivityLog>, AxiosError>
 ): responseInterface<PaginatedResult<ActivityLog>, AxiosError> => {
-    const uuid = ServerContext.useStoreState((state) => state.server.data!.uuid);
+    const uuid = ServerContext.useStoreState((state) => state.server.data?.uuid);
+    const key = useServerSWRKey(['activity', useFilteredObject(filters || {})]);
 
     return useSWR<PaginatedResult<ActivityLog>>(
-        ['server:activty', uuid, useFilteredObject(filters || {})],
+        key,
         async () => {
             const { data } = await http.get(`/api/client/servers/${uuid}/activity`, {
                 params: {
